Cache table installation state to avoid duplicate queries (#74)
* Cache table installation state to avoid duplicate queries

When 'SHOW TABLES LIKE ...' succeeds (i.e. WPML is installed), set a
cache value 'installed' to true. This means that once WPML is installed
to the database, there will be no more repeated 'SHOW TABLES LIKE ..'
queries (previously: multiple times each page request) executed against
the database.

The version checks are left untouched, they should still be used as-is
in conjunction with `isInstalled()`. This only caches the installation
state of the tables in the database (for any version) to optimize the
common case of "yes, they're installed."

The cache variable is unset on uninstall so this should play nice with
tests and with users uninstalling then wishing to reinstall the plugin
within the default cache variable expiry window (1 hour).

Closes #71.